Renommage commun, controller maj, suppression service upload.

+++ src/main/java/io/swagger/api/ImageApiController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.RestController;
 import io.swagger.annotations.ApiParam;
 import io.swagger.model.Image;
 import io.swagger.repository.ImageRepository;
-import io.swagger.service.image.ImageService;
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-01-16T20:34:37.063Z")
 
 @RestController
@@ -23,20 +22,23 @@ public class ImageApiController implements ImageApi {
 	
 	@Autowired
 	private ImageRepository repo;
-	
-	@Autowired
-	private ImageService imgServ;
 
     public ResponseEntity<Image> createImageUsingPOST(@ApiParam(value = "image" ,required=true )  @Valid @RequestBody Image image) {
         // do some magic!
-    	repo.save(image);
-        return new ResponseEntity<Image>(HttpStatus.OK);
+    	image.setId(null);
+    	image = repo.save(image);
+        return new ResponseEntity<Image>(image, HttpStatus.OK);
     }
 
     public ResponseEntity<Void> deleteImageUsingDELETE(@ApiParam(value = "id",required=true ) @PathVariable("id") Long id) {
         // do some magic!
+    	if(repo.exists(id))
+    	{
     	repo.delete(id);
         return new ResponseEntity<Void>(HttpStatus.OK);
+    	}
+    	else
+    		return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
     }
 
     public ResponseEntity<List<Image>> getAllImagesUsingGET() {
@@ -48,15 +50,24 @@ public class ImageApiController implements ImageApi {
     public ResponseEntity<Image> getImageUsingGET(@ApiParam(value = "id",required=true ) @PathVariable("id") Long id) {
         // do some magic!
     	Image image = new Image();
-    	image = repo.findOne(id);	
-        return new ResponseEntity<Image>(image,HttpStatus.OK);
+    	image = repo.findOne(id);
+    	if(image == null)
+    		return new ResponseEntity<Image>(HttpStatus.NOT_FOUND);
+    	else
+    		return new ResponseEntity<Image>(image,HttpStatus.OK);
     }
 
     public ResponseEntity<Image> updateImageUsingPUT(@ApiParam(value = "id",required=true ) @PathVariable("id") Long id, @Valid @RequestBody Image image) {
         // do some magic!
     	Image img = repo.findOne(id);
-    	img = repo.saveAndFlush(image);
-        return new ResponseEntity<Image>(HttpStatus.OK);
+    	if(img == null)
+    		return new ResponseEntity<Image>(HttpStatus.NOT_FOUND);
+    	else
+	    	{
+    		image.setId(img.getId());
+	    	img = repo.saveAndFlush(image);
+	        return new ResponseEntity<Image>(img, HttpStatus.OK);
+    	}
     }
 
 }
